3d_design: separate package installation from container instantiation.
This significantly speeds up instantiation time. The downside is the image remains cached and must be pre-generated.

+++ 3d_design/Run_Interactively.sh
@@ -1,7 +1,5 @@
 #!/bin/bash
 
-set -eu
-
 # This script launches a Docker image that can pass-through X11 and OpenGL windows.
 #
 # It might fail! Especially if .Xauthority files are used.
@@ -13,56 +11,47 @@ set -eu
 # See https://github.com/jessfraz/dockerfiles/issues/359 for info about the MIT-SHM (X_ShmPutImage) error fix.
 #
 
-# Create a script that can be run from inside the container.
+set -eu
+
+if [ "$EUID" -eq 0 ] ; then
+    printf "This script should not be run as root. Refusing to continue.\n" 1>&2
+    exit 1
+fi
+
+# Create a simple 'init' script that can be run from inside the container.
 #
-# All arguments sent to the present script are forwarded to the internal script.
+# All arguments sent to the present script are forwarded to this internal script.
 #
 internal_run_script="$(mktemp)"
 trap "{ rm '${internal_run_script}' ; }" EXIT  # Remove the temp file on exit.
 
-# The calling user's username, which we will re-use inside the container.
+# The calling user's username on the host, which we will re-use inside the container.
 uname=$(id -u -n)
-uid=$(id -u)
 gname=$(id -g -n)
+uid=$(id -u)
 gid=$(id -g)
 
 cat > "${internal_run_script}" <<EOF
 #!/bin/bash
 
+# Copy host files to facilitate easier interoperation.
 #cp /etc/passwd_prototype /etc/passwd
 #cp /etc/group_prototype  /etc/group
 #cp /etc/shadow_prototype /etc/shadow
-#adduser --force-badname --disabled-password --gecos '' --quiet '_apt'
+
+# Create a user with the same credentials as the host user.
 groupadd -g ${gid} -f ${gname} || true
 useradd -g ${gname} -u ${uid} -G video -m -d '/home/container_${uname}' -s /bin/bash ${uname}
+
+# Copy host Xauth file.
 cp /etc/Xauthority_prototype /home/container_${uname}/.Xauthority
 chown ${uname}:${gname} /home/container_${uname}/.Xauthority
 
-# The following will enable OpenGL acceleration inside a docker container.
-#
-# Note: The following seems to NOT be needed for OpenGL with SFML.
-#
+# Install additional packages.
+#apt-get -y update
 #apt-get -y install firefox-esr
-#apt-get -y install mesa-va-drivers mesa-vdpau-drivers va-driver-all vdpau-driver-all libva-drm2 libva-x11-2 libva2 libvdpau-va-gl1 libvdpau1 
-#apt-get -y install va-driver-all vdpau-driver-all libva-drm2 libva-x11-2 libva2 libvdpau-va-gl1 libvdpau1 
-apt-get -y update
-apt-get -y install inkscape freecad openscad cura gimp vim rsync gedit
-#apt-get -y install gedit
 
-# Copy the portable_dcma binaries into the container.
-#rsync -r /root/portable_dcma/ /usr/bin/
-
-# Remove the existing dcma binary to avoid confusion.
-#rm /usr/bin/dicomautomaton_dispatcher
-
-# Add a user to use X.
 #
-# Note: Need to use a non-root user.
-#
-#useradd --system -g users -G video -m -s /bin/bash auser 2>/dev/null
-#groupadd -f -o -g 1000 users
-#useradd -g users -u 1000 -G video -m -s /bin/bash auser
-
 printf 'export PS1="${uname}@container> "\n' >> /home/${uname}/.bashrc
 
 # Pass-through any arguments sent to the run script (not this script, the script that created this script).
@@ -73,7 +62,6 @@ if [ -z "\$all_args" ] ; then
     runuser -u ${uname} -- /bin/bash -i
 
 else # Pass the command through.
-    #runuser -u ${uname} -- /home/${uname}/portable_dcma/portable_dcma /tmp/CT*dcm
     runuser -u ${uname} -- \$all_args
 
 fi
@@ -94,31 +82,24 @@ sudo \
     `# Map X-related host locations into the container. ` \
     `# -v /tmp/.X11-unix:/tmp/.X11-unix:rw ` \
     -v /tmp/:/tmp/:rw \
-    -v "${internal_run_script}":/x11_launch_script.sh:ro \
+    -v "$HOME"/.Xauthority:/etc/Xauthority_prototype:ro \
     -v /dev/null:/dev/input/js0:ro \
     -v /dev/null:/dev/input/js1:ro \
     -v /dev/null:/dev/input/js2:ro \
+    -v "${internal_run_script}":/x11_launch_script.sh:ro \
     \
-    `# Map all host users into the container. ` \
-    -v /etc/passwd:/etc/passwd_prototype:ro \
-    -v /etc/group:/etc/group_prototype:ro \
-    -v /etc/shadow:/etc/shadow_prototype:ro \
-    -v "$HOME"/.Xauthority:/etc/Xauthority_prototype:ro \
-    \
-    `# Map the DCMA binary locations into the container. ` \
-    -v /home/hal/portable_dcma/:/root/portable_dcma/:ro \
-    -v /home/hal/portable_dcma/:/home/${uname}/portable_dcma/:ro \
-    \
-    `# Map various locations from host into the container. ` \
+    `# Map various locations from host into the container prospectively. ` \
     -v /media/:/media/:ro \
     -v /media/sf_U_DRIVE/Profile/Desktop:/home/${uname}/Desktop/:ro \
-    -v /mnt/scratch_A/Research_Archives:/home/${uname}/Research_Archives/:rw \
     -v /:/host_root/:ro \
-    \
     -v "$(pwd)":/start/:rw \
     -w /start/ \
     \
-    debian:stable \
+    3d_design:latest \
     \
     /x11_launch_script.sh
 
+#    debian:stable \
+#    -v /etc/passwd:/etc/passwd_prototype:ro \
+#    -v /etc/group:/etc/group_prototype:ro \
+#    -v /etc/shadow:/etc/shadow_prototype:ro \
